+++ amd/comgr/src/comgr-compiler.h
@@ -143,6 +143,7 @@ public:
   amd_comgr_status_t preprocessToSource();
   amd_comgr_status_t compileToBitcode(bool WithDeviceLibs = false);
   amd_comgr_status_t compileToRelocatable();
+  amd_comgr_status_t unbundle();
   amd_comgr_status_t linkBitcodeToBitcode();
   amd_comgr_status_t codeGenBitcodeToRelocatable();
   amd_comgr_status_t codeGenBitcodeToAssembly();