+++ amd/comgr/test/unbundle_hip_test.c
@@ -31,37 +31,40 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS WITH
  * THE SOFTWARE.
  *
- *******************************************************************************/
+ ******************************************************************************/
 
 /// -------
 //  Manual recreation of Comgr bundle linking
 //
-//    // Create bundles
-//    clang -c --offload-arch=gfx906 -emit-llvm -fgpu-rdc \
-//    --gpu-bundle-output square.hip double.hip cube.hip
+//    // Create bitcode bundles
+//    clang -c --offload-arch=gfx900 -emit-llvm -fgpu-rdc \
+//    --gpu-bundle-output square.hip cube.hip
 //
+//    // Create object file bundles
+//    clang -c --offload-arch=gfx900 --gpu-bundle-output \
+//    double.hip
+//
+//    // Create archive bundle
 //    llvm-ar rc cube.a cube.bc
 //
-//    // Manually unbundle
+//    // Manually unbundle bitcode bundle
 //    clang-offload-bundler -type=bc \
-//    -targets=hip-amdgcn-amd-amdhsa-gfx906 \
-//    -input=square.bc -output=square-gfx906.bc \
+//    -targets=hip-amdgcn-amd-amdhsa-gfx900 \
+//    -input=square.bc -output=square-gfx900.bc \
 //    -unbundle -allow-missing-bundles
 //
-//    clang-offload-bundler -type=bc \
-//    -targets=hip-amdgcn-amd-amdhsa-gfx906 \
-//    -input=double.bc -output=double-gfx906.bc \
+//    // Manually unbundle object file bundle
+//    clang-offload-bundler -type=o \
+//    -targets=hip-amdgcn-amd-amdhsa-gfx900 \
+//    -input=double.o -output=double-gfx900.o \
 //    -unbundle -allow-missing-bundles
 //
+//    // Manually unbundle archive bundle
 //    clang-offload-bundler -type=a \
-//    -targets=hip-amdgcn-amd-amdhsa-gfx906 \
-//    -input=cube.a -output=cube-gfx906.a \
+//    -targets=hip-amdgcn-amd-amdhsa-gfx900 \
+//    -input=cube.a -output=cube-gfx900.a \
 //    -unbundle -allow-missing-bundles \
 //    -hip-openmp-compatible
-//
-//    // Manually link
-//    llvm-link square-gfx906.bc double-gfx906.bc cube-gfx906.a \
-//    -o gold/gold-linked-bitcode-gfx906.bc
 
 #include "amd_comgr.h"
 #include "common.h"
@@ -70,42 +73,42 @@
 #include <string.h>
 
 int main(int Argc, char *Argv[]) {
-  char *BufArchive, *BufBitcode1, *BufBitcode2;
-  size_t SizeArchive, SizeBitcode1, SizeBitcode2;
-  amd_comgr_data_t DataArchive, DataBitcode1, DataBitcode2;
-  amd_comgr_data_set_t DataSetBundled, DataSetLinked, DataSetReloc,
+  char *BufBitcode, *BufObjectFile, *BufArchive;
+  size_t SizeBitcode, SizeObjectFile, SizeArchive;
+  amd_comgr_data_t DataBitcode, DataObjectFile, DataArchive;
+  amd_comgr_data_set_t DataSetBundled, DataSetUnbundled,
+                       DataSetLinked, DataSetReloc,
                        DataSetExec;
-  amd_comgr_action_info_t ActionInfo;
+  amd_comgr_action_info_t ActionInfoUnbundle, ActionInfoLink;
   amd_comgr_status_t Status;
 
-  const char *IsaName = "amdgcn-amd-amdhsa--gfx906";
-
-  SizeBitcode1 = setBuf("./source/square.bc", &BufBitcode1);
-  SizeBitcode2 = setBuf("./source/double.bc", &BufBitcode2);
+  SizeBitcode = setBuf("./source/square.bc", &BufBitcode);
+  SizeObjectFile = setBuf("./source/double.o", &BufObjectFile);
   SizeArchive = setBuf("./source/cube.a", &BufArchive);
 
   // Create Bundled dataset
   Status = amd_comgr_create_data_set(&DataSetBundled);
   checkError(Status, "amd_comgr_create_data_set");
 
-  // Bitcode1
-  Status = amd_comgr_create_data(AMD_COMGR_DATA_KIND_BC_BUNDLE, &DataBitcode1);
+  // Bitcode
+  Status = amd_comgr_create_data(AMD_COMGR_DATA_KIND_BC_BUNDLE, &DataBitcode);
   checkError(Status, "amd_comgr_create_data");
-  Status = amd_comgr_set_data(DataBitcode1, SizeBitcode1, BufBitcode1);
+  Status = amd_comgr_set_data(DataBitcode, SizeBitcode, BufBitcode);
   checkError(Status, "amd_comgr_set_data");
-  Status = amd_comgr_set_data_name(DataBitcode1, "square");
+  Status = amd_comgr_set_data_name(DataBitcode, "square");
   checkError(Status, "amd_comgr_set_data_name");
-  Status = amd_comgr_data_set_add(DataSetBundled, DataBitcode1);
+  Status = amd_comgr_data_set_add(DataSetBundled, DataBitcode);
   checkError(Status, "amd_comgr_data_set_add");
 
-  // Bitcode2
-  Status = amd_comgr_create_data(AMD_COMGR_DATA_KIND_BC_BUNDLE, &DataBitcode2);
+  // ObjectFile
+  Status = amd_comgr_create_data(AMD_COMGR_DATA_KIND_OBJ_BUNDLE,
+                                 &DataObjectFile);
   checkError(Status, "amd_comgr_create_data");
-  Status = amd_comgr_set_data(DataBitcode2, SizeBitcode2, BufBitcode2);
+  Status = amd_comgr_set_data(DataObjectFile, SizeObjectFile, BufObjectFile);
   checkError(Status, "amd_comgr_set_data");
-  Status = amd_comgr_set_data_name(DataBitcode2, ""); // test blank name
+  Status = amd_comgr_set_data_name(DataObjectFile, ""); // test blank name
   checkError(Status, "amd_comgr_set_data_name");
-  Status = amd_comgr_data_set_add(DataSetBundled, DataBitcode2);
+  Status = amd_comgr_data_set_add(DataSetBundled, DataObjectFile);
   checkError(Status, "amd_comgr_data_set_add");
 
   // Archive
@@ -118,81 +121,176 @@ int main(int Argc, char *Argv[]) {
   Status = amd_comgr_data_set_add(DataSetBundled, DataArchive);
   checkError(Status, "amd_comgr_data_set_add");
 
-  // Set up ActionInfo
-  Status = amd_comgr_create_action_info(&ActionInfo);
-  checkError(Status, "amd_comgr_create_action_info");
+  // Unbundle explicitly via UNBUNDLE action
+  {
+    // Set up ActionInfo
+    Status = amd_comgr_create_action_info(&ActionInfoUnbundle);
+    checkError(Status, "amd_comgr_create_action_info");
 
-  Status =
-      amd_comgr_action_info_set_language(ActionInfo, AMD_COMGR_LANGUAGE_HIP);
-  checkError(Status, "amd_comgr_action_info_set_language");
+    Status =
+      amd_comgr_action_info_set_language(ActionInfoUnbundle,
+                                         AMD_COMGR_LANGUAGE_HIP);
+    checkError(Status, "amd_comgr_action_info_set_language");
 
-  Status = amd_comgr_action_info_set_isa_name(ActionInfo, IsaName);
+    const char *BundleEntryIDs[] = {"host-x86_64-unknown-linux-gnu",
+      "hip-amdgcn-amd-amdhsa-gfx900"};
+    size_t BundleEntryIDsCount =
+      sizeof(BundleEntryIDs) / sizeof(BundleEntryIDs[0]);
+    Status = amd_comgr_action_info_set_bundle_entry_ids(ActionInfoUnbundle,
+                                                        BundleEntryIDs, 2);
 
-  // Unbundle
-  Status = amd_comgr_create_data_set(&DataSetLinked);
-  checkError(Status, "amd_comgr_create_data_set");
-  Status = amd_comgr_do_action(AMD_COMGR_ACTION_LINK_BC_TO_BC, ActionInfo,
-                               DataSetBundled, DataSetLinked);
-  checkError(Status, "amd_comgr_do_action");
+    // Unbundle
+    Status = amd_comgr_create_data_set(&DataSetUnbundled);
+    checkError(Status, "amd_comgr_create_data_set");
+    Status = amd_comgr_do_action(AMD_COMGR_ACTION_UNBUNDLE, ActionInfoUnbundle,
+                                 DataSetBundled, DataSetUnbundled);
+    checkError(Status, "amd_comgr_do_action");
 
-  // Check Linked bitcode count
-  size_t Count;
-  Status = amd_comgr_action_data_count(DataSetLinked,
-                                       AMD_COMGR_DATA_KIND_BC, &Count);
-  checkError(Status, "amd_comgr_action_data_count");
+    // Check Bitcode count
+    size_t Count;
+    Status = amd_comgr_action_data_count(DataSetUnbundled,
+                                         AMD_COMGR_DATA_KIND_BC, &Count);
+    checkError(Status, "amd_comgr_action_data_count");
 
-  if (Count != 1) {
-    printf("Bundled bitcode linking: "
-           "produced %zu bitcodes (expected 1)\n",
-           Count);
-    exit(1);
+    if (Count != 2) {
+      printf("Unbundling: "
+             "produced %zu bitcodes (expected 2)\n",
+             Count);
+      exit(1);
+    }
+
+    // Check ObjectFile count
+    Status = amd_comgr_action_data_count(DataSetUnbundled,
+                                         AMD_COMGR_DATA_KIND_EXECUTABLE,
+                                         &Count);
+    checkError(Status, "amd_comgr_action_data_count");
+
+    if (Count != 2) {
+      printf("Unbundling: "
+             "produced %zu object files (expected 2)\n",
+             Count);
+      exit(1);
+    }
+
+    // Check Archive count
+    Status = amd_comgr_action_data_count(DataSetUnbundled,
+                                         AMD_COMGR_DATA_KIND_AR, &Count);
+    checkError(Status, "amd_comgr_action_data_count");
+
+    if (Count != 2) {
+      printf("Unbundle: "
+             "produced %zu archives (expected 2)\n",
+             Count);
+      exit(1);
+    }
+
+    // Check Bundle Entry IDs
+    size_t Size;
+    char **ComgrBundleEntryIDs;
+
+    Status = amd_comgr_action_info_get_bundle_entry_ids(ActionInfoUnbundle,
+                                                        &Size, NULL);
+    ComgrBundleEntryIDs = malloc(Size * sizeof(char));
+
+    checkError(Status, "amd_comgr_action_info_get_bundle_entry_ids");
+    Status = amd_comgr_action_info_get_bundle_entry_ids(ActionInfoUnbundle,
+                                                        &Size,
+                                                        ComgrBundleEntryIDs);
+
+    for (int i = 0; i < 2; i++) {
+      if (strcmp(ComgrBundleEntryIDs[i], BundleEntryIDs[i])) {
+        printf("BundleEntryID mismatch. Expected \"%s\", returned \"%s\"\n",
+               BundleEntryIDs[i],
+               ComgrBundleEntryIDs[i]);
+        checkError(AMD_COMGR_STATUS_ERROR,
+                   "amd_comgr_action_info_get_bundle_entry_ids");
+      }
+    }
+
+    free(ComgrBundleEntryIDs);
   }
 
-  // Compile to relocatable
-  Status = amd_comgr_create_data_set(&DataSetReloc);
-  checkError(Status, "amd_comgr_create_data_set");
+  // Unbundle silently via LINK action
+  {
+    // Set up ActionInfo
+    Status = amd_comgr_create_action_info(&ActionInfoLink);
+    checkError(Status, "amd_comgr_create_action_info");
 
-  Status = amd_comgr_do_action(AMD_COMGR_ACTION_CODEGEN_BC_TO_RELOCATABLE,
-                               ActionInfo, DataSetLinked, DataSetReloc);
-  checkError(Status, "amd_comgr_do_action");
+    Status =
+      amd_comgr_action_info_set_language(ActionInfoLink,
+                                         AMD_COMGR_LANGUAGE_HIP);
+    checkError(Status, "amd_comgr_action_info_set_language");
 
-  Status = amd_comgr_action_data_count(DataSetReloc,
-                                       AMD_COMGR_DATA_KIND_RELOCATABLE, &Count);
-  checkError(Status, "amd_comgr_action_data_count");
+    const char *IsaName = "amdgcn-amd-amdhsa--gfx900";
+    Status = amd_comgr_action_info_set_isa_name(ActionInfoLink, IsaName);
 
-  if (Count != 1) {
-    printf("AMD_COMGR_ACTION_CODEGEN_BC_TO_RELOCATABLE Failed: "
-           "produced %zu source objects (expected 1)\n",
-           Count);
-    exit(1);
-  }
+    // Unbundle
+    Status = amd_comgr_create_data_set(&DataSetLinked);
+    checkError(Status, "amd_comgr_create_data_set");
+    Status = amd_comgr_do_action(AMD_COMGR_ACTION_LINK_BC_TO_BC, ActionInfoLink,
+                                 DataSetBundled, DataSetLinked);
+    checkError(Status, "amd_comgr_do_action");
 
-  // Compile to executable
-  Status = amd_comgr_create_data_set(&DataSetExec);
-  checkError(Status, "amd_comgr_create_data_set");
+    // Check Linked bitcode count
+    size_t Count;
+    Status = amd_comgr_action_data_count(DataSetLinked,
+                                         AMD_COMGR_DATA_KIND_BC, &Count);
+    checkError(Status, "amd_comgr_action_data_count");
 
-  Status = amd_comgr_action_info_set_option_list(ActionInfo, NULL, 0);
-  checkError(Status, "amd_comgr_action_info_set_option_list");
+    if (Count != 1) {
+      printf("Bundled bitcode linking: "
+             "produced %zu bitcodes (expected 1)\n",
+             Count);
+      exit(1);
+    }
 
-  Status = amd_comgr_do_action(AMD_COMGR_ACTION_LINK_RELOCATABLE_TO_EXECUTABLE,
-                               ActionInfo, DataSetReloc, DataSetExec);
-  checkError(Status, "amd_comgr_do_action");
+    // Compile to relocatable
+    Status = amd_comgr_create_data_set(&DataSetReloc);
+    checkError(Status, "amd_comgr_create_data_set");
 
-  Status = amd_comgr_action_data_count(DataSetExec,
-                                       AMD_COMGR_DATA_KIND_EXECUTABLE, &Count);
-  checkError(Status, "amd_comgr_action_data_count");
+    Status = amd_comgr_do_action(AMD_COMGR_ACTION_CODEGEN_BC_TO_RELOCATABLE,
+                                 ActionInfoLink, DataSetLinked, DataSetReloc);
+    checkError(Status, "amd_comgr_do_action");
 
-  if (Count != 1) {
-    printf("AMD_COMGR_ACTION_LINK_RELOCATABLE_TO_EXECUTABLE Failed: "
-           "produced %zu executable objects (expected 1)\n",
-           Count);
-    exit(1);
+    Status = amd_comgr_action_data_count(DataSetReloc,
+                                         AMD_COMGR_DATA_KIND_RELOCATABLE, &Count);
+    checkError(Status, "amd_comgr_action_data_count");
+
+    if (Count != 1) {
+      printf("AMD_COMGR_ACTION_CODEGEN_BC_TO_RELOCATABLE Failed: "
+             "produced %zu source objects (expected 1)\n",
+             Count);
+      exit(1);
+    }
+
+    // Compile to executable
+    Status = amd_comgr_create_data_set(&DataSetExec);
+    checkError(Status, "amd_comgr_create_data_set");
+
+    Status = amd_comgr_do_action(AMD_COMGR_ACTION_LINK_RELOCATABLE_TO_EXECUTABLE,
+                                 ActionInfoLink, DataSetReloc, DataSetExec);
+    checkError(Status, "amd_comgr_do_action");
+
+    Status = amd_comgr_action_data_count(DataSetExec,
+                                         AMD_COMGR_DATA_KIND_EXECUTABLE, &Count);
+    checkError(Status, "amd_comgr_action_data_count");
+
+    if (Count != 1) {
+      printf("AMD_COMGR_ACTION_LINK_RELOCATABLE_TO_EXECUTABLE Failed: "
+             "produced %zu executable objects (expected 1)\n",
+             Count);
+      exit(1);
+    }
   }
 
   // Cleanup
-  Status = amd_comgr_destroy_action_info(ActionInfo);
+  Status = amd_comgr_destroy_action_info(ActionInfoUnbundle);
+  checkError(Status, "amd_comgr_destroy_action_info");
+  Status = amd_comgr_destroy_action_info(ActionInfoLink);
   checkError(Status, "amd_comgr_destroy_action_info");
   Status = amd_comgr_destroy_data_set(DataSetBundled);
+  checkError(Status, "amd_comgr_destroy_data_set");
+  Status = amd_comgr_destroy_data_set(DataSetUnbundled);
   checkError(Status, "amd_comgr_destroy_data_set");
   Status = amd_comgr_destroy_data_set(DataSetLinked);
   checkError(Status, "amd_comgr_destroy_data_set");
@@ -200,14 +298,16 @@ int main(int Argc, char *Argv[]) {
   checkError(Status, "amd_comgr_destroy_data_set");
   Status = amd_comgr_destroy_data_set(DataSetExec);
   checkError(Status, "amd_comgr_destroy_data_set");
-  Status = amd_comgr_release_data(DataBitcode1);
+  Status = amd_comgr_release_data(DataBitcode);
   checkError(Status, "amd_comgr_release_data");
-  Status = amd_comgr_release_data(DataBitcode2);
+  Status = amd_comgr_release_data(DataObjectFile);
   checkError(Status, "amd_comgr_release_data");
   Status = amd_comgr_release_data(DataArchive);
   checkError(Status, "amd_comgr_release_data");
 
-  free(BufBitcode1);
-  free(BufBitcode2);
+  free(BufBitcode);
+  free(BufObjectFile);
   free(BufArchive);
+
+  return 0;
 }